ssh transport mod request
There must be an option to skip disabling StrictHostKeyChecking. Kitchen ssh transport doesn't anyhow check if it connects to right machine or not. We may execute let's ay  'kitchen verify' command on some stale data saved in '.kitchen/*' directory. If somebody uses ssh key auth for every of his virtual machines - possible situation if kitchen transfers data and executes commands on some foreign machine recreated with the same IP. We have to have a way to prevent such a problems.